First commit, first lesson since re repo

+++ Intro to opengl and freeflut/Structures.h
@@ -34,5 +34,5 @@ struct Mesh {
     Vertex* Vertices;
     Color* Colors;
     GLushort* Indices;
-    int VertexCount, ColorCount, IndexCount;
+    GLfloat VertexCount, ColorCount, IndexCount;
 };
